requirements.txt
flask
openai
mcp
python-dotenv
